Pulling registry and interfaces up, makePublicInterfaces() base.
